% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceptanceExt.R
\name{acceptanceExt}
\alias{acceptanceExt}
\title{Calculate acceptance probability}
\usage{
acceptanceExt(x.current, x.proposal, ldelta, w, Model, lambda = NULL)
}
\arguments{
\item{x.current}{Current state of MCMC}

\item{x.proposal}{Proposed new state of the MCMC}

\item{ldelta}{Combined normalizing constant and delta}

\item{w}{Weight put on "outsideness"}

\item{Model}{Type of model ("uniform" or "poisson")}

\item{lambda}{Rate parameters for "poisson" Model}
}
\description{
Calculate acceptance probability
}
