% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateAlpha.R
\name{calculateAlpha}
\alias{calculateAlpha}
\title{Calculate Acceptance ratio}
\usage{
calculateAlpha(xCurrent, xProposal, loggamma, w = 0, dist = "Unif")
}
\arguments{
\item{xCurrent}{Current point}

\item{xProposal}{Proposed point}

\item{loggamma}{log(delta) used to define epsilon}

\item{w}{Weight of negativity}

\item{dist}{Model distribution}
}
\value{
Returns the Acceptance ratio
}
\description{
Calculate Acceptance ratio
}
