% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_transition_matrices.R
\name{compute_transition_matrices}
\alias{compute_transition_matrices}
\title{Computes transition matrices for all moves}
\usage{
compute_transition_matrices(fibre, B, epsilon = 1e-04)
}
\arguments{
\item{fibre}{A matrix that contains all every point as column}

\item{B}{A matrix that contains everey move of a basis as columns}

\item{epsilon}{Tolerance to off-integerness}
}
\value{
Returns a list of matrices. Each is the transition matrix for a particular move of B. These can be used to find the overall transition matrix for the whole setup.
}
\description{
Computes transition matrices for all moves
}
