% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCPLB.R
\name{computeCPLB}
\alias{computeCPLB}
\title{Computes a column partition lattice basis}
\usage{
computeCPLB(A, A2Idx)
}
\arguments{
\item{A}{Configuration matrix}

\item{A2Idx}{Index of columns of A2}
}
\value{
Returns a column partition lattice basis as matrix
}
\description{
Computes a column partition lattice basis
}
