% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeMESS.R
\name{computeMESS}
\alias{computeMESS}
\title{Compute the multidimensional ESS for the sample}
\usage{
computeMESS(x, A)
}
\arguments{
\item{x}{Sample generated by LIPS (matrix)}

\item{A}{Configuration matrix}
}
\description{
Compute the multidimensional ESS for the sample
}
