% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contingencyTable}
\alias{contingencyTable}
\title{3-way contingency table: Book-crossing data}
\format{
\subsection{\code{contingencyTable}}{

A list containing all that is needed to run the examples of the paper:
\describe{
\item{A}{Configuration Matrix}
\item{y}{Observation vector}
\item{lambda}{Poisson rates for the cells of the table}
\item{LatticeBasis}{Lattice basis found using the column partition as discussed in Extension Sampler chapter}
\item{BasicMoves}{Collection of moves that forms a 1-Markov basis for this example}
\item{MarkovBasis}{A Markov basis that contains the lattice basis and the collection of basic moves.}
}
}
}
\source{
<doi: 10.1093/biomet/asaa083>
}
\usage{
contingencyTable
}
\description{
Data that was used in Hazelton et al. (2021) and collected by Ziegler et al. (2005)
}
\keyword{datasets}
