% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensionSampler.R
\name{extensionSampler}
\alias{extensionSampler}
\title{Extension Samplers}
\usage{
extensionSampler(
  A,
  y,
  B = NULL,
  Model = "uniform",
  lambda = NULL,
  extension = "p",
  p = 0,
  A2Idx = NULL,
  a = NULL,
  b = 0,
  loggamma = 0,
  w = 0,
  xStart = NULL,
  nSample = 1e+05,
  nBurnin = 10000,
  nChains = 4,
  thinning = 1
)
}
\arguments{
\item{A}{Configuration Matrix}

\item{y}{(Corrupted/aggregated) Observation Vector}

\item{B}{Set of moves used (p-Markov, CPLB or Full Markov)}

\item{Model}{Type of model ("uniform" or "poisson")}

\item{lambda}{Rate parameters for "poisson" Model}

\item{extension}{Type of extension ("p", "hyperrectangle", "knapsack")}

\item{p}{Relaxation limit}

\item{A2Idx}{Free variables correponding to matrix A_2}

\item{a}{"hyperrectangle" extension: bounds on hyperrectangle, "knapsack": objective function}

\item{b}{Upper bound of objective function for knapsack extension}

\item{loggamma}{Combined normalizing constant and delta}

\item{w}{Weight put on "outsideness"}

\item{xStart}{Starting points for each chain. Needs to be fed in as a matrix with number of rows = number of chains}

\item{nSample}{Number of samples}

\item{nBurnin}{Number of burnin steps}

\item{nChains}{Number of chains}

\item{thinning}{Thinning parameter}
}
\description{
Extension Samplers
}
