% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_fibre.R
\name{generate_fibre}
\alias{generate_fibre}
\title{Compute a specific fibre}
\usage{
generate_fibre(A, y, epsilon = 1e-04)
}
\arguments{
\item{A}{Configuration matrix}

\item{y}{Observation vector}

\item{epsilon}{Tolerance for off-integerness}
}
\value{
Returns a Matrix containing all points of the fibre as columns.
}
\description{
Compute a specific fibre
}
