% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_fibre_extension.R
\name{generate_fibre_extension}
\alias{generate_fibre_extension}
\title{Compute a specific p-fibre extension}
\usage{
generate_fibre_extension(A, y, p, epsilon = 1e-04)
}
\arguments{
\item{A}{Configuration matrix}

\item{y}{Observation vector}

\item{p}{Relaxation limit}

\item{epsilon}{Tolerance for off-integerness}
}
\value{
Returns a Matrix containing all points of the fibre as columns.
}
\description{
Compute a specific p-fibre extension
}
