% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{highway}
\alias{highway}
\title{Network tomography example: A6 around Leicester}
\format{
\subsection{\code{highway}}{

A list containing all that is needed to run the examples of the paper:
\describe{
\item{A}{Configuration Matrix}
\item{y}{Observation vector}
\item{lambda}{Traffic rates for the different routes through A6}
\item{LatticeBasis}{Lattice basis found using the column partition k^(-1)(1) = 1,2,3,4,5,6,7}
\item{MarkovBasis}{A Markov basis that contains the lattice basis used for the extension samplers}
}
}
}
\source{
<doi: 10.1093/biomet/asaa083>
}
\usage{
highway
}
\description{
Data that was used in Hazelton et al. (2021) about dymanical lattice basis samplers
}
\keyword{datasets}
