% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsackSampler.R
\name{knapsackSampler}
\alias{knapsackSampler}
\title{knapsack-Samplers}
\usage{
knapsackSampler(
  A,
  y,
  B = NULL,
  Model = "uniform",
  lambda = NULL,
  CPLBIdx = NULL,
  a = NULL,
  b = 0,
  ldelta = 0,
  w = 0,
  x.start = NULL,
  n.sample = 1e+05,
  n.burnin = 10000,
  chain.id = 1,
  thinning = 1
)
}
\arguments{
\item{A}{Configuration Matrix}

\item{y}{(Corrupted/aggregated) Observation Vector}

\item{B}{Set of moves used (p-Markov, CPLB or Full Markov)}

\item{Model}{Type of model ("uniform" or "poisson")}

\item{lambda}{Rate parameters for "poisson" Model}

\item{CPLBIdx}{Free variables correponding to matrix A_2}

\item{a}{"hyperrectangle" extension: bounds on hyperrectangle, "knapsack": objective function}

\item{b}{Upper bound of objective function for knapsack extension}

\item{ldelta}{Combined normalizing constant and delta}

\item{w}{Weight put on "outsideness"}

\item{x.start}{Starting points for each chain. Needs to be fed in as a matrix with number of rows = number of chains}

\item{n.sample}{Number of samples}

\item{n.burnin}{Number of burnin steps}

\item{chain.id}{Which number chain?}

\item{thinning}{Thinning parameter}
}
\description{
knapsack-Samplers
}
