% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onlyPositive.R
\name{onlyPositive}
\alias{onlyPositive}
\title{Replace negative rows with NA}
\usage{
onlyPositive(x)
}
\arguments{
\item{x}{Sample generated by LIPS (matrix)}
}
\description{
Replace negative rows with NA
}
