% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_sampler.R
\name{p_sampler}
\alias{p_sampler}
\title{Sampler over the B_p-extension}
\usage{
p_sampler(
  A,
  y,
  n.samples = 1e+05,
  n.chains = 4,
  n.burnin = 10000,
  B = NULL,
  CPLB_idx = NULL,
  x.start = NULL,
  p = 0,
  ldelta = 0,
  w = 0,
  dist = "unif"
)
}
\arguments{
\item{A}{Configuration matrix}

\item{y}{Observation vector}

\item{n.samples}{Number of samples per chain}

\item{n.chains}{Number of chains}

\item{n.burnin}{length of burnin}

\item{B}{A set of vectors connecting the M_p-extension}

\item{CPLB_idx}{A set indicating A_2 columns}

\item{x.start}{Initial point}

\item{p}{Relaxation limit}

\item{ldelta}{log(delta) determining epsilon}

\item{w}{Negativity weight}

\item{dist}{Model distribution}
}
\value{
Returns a matrix containing the samples over the M_p-extension generated by a hit-and-run algorithm
}
\description{
Sampler over the B_p-extension
}
