% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposePoint_hypercube_gibbs.R
\name{proposePoint_hypercube_gibbs}
\alias{proposePoint_hypercube_gibbs}
\title{Propose a new point in the hypercube extension (Gibbs)}
\usage{
proposePoint_hypercube_gibbs(x, B, ldelta, w, CPLB_idx, a, dist)
}
\arguments{
\item{x}{Current point}

\item{B}{Sampling directions (Should be a column partition lattice basis)'}

\item{ldelta}{log(delta) as by definition of this sampler and extension}

\item{w}{Weight in exponential as described in definition of extension distribution}

\item{CPLB_idx}{Index of the A_2 matrix}

\item{a}{Bounds of the hypercuboid}

\item{dist}{Proposed model}
}
\value{
Returns a new proposed point in the hypercube extension
}
\description{
Propose a new point in the hypercube extension (Gibbs)
}
