% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pSampler.R
\name{pSampler}
\alias{pSampler}
\title{p-samplers}
\usage{
pSampler(
  A,
  y,
  B,
  Model = "uniform",
  lambda = NULL,
  p = 0,
  ldelta = 0,
  w = 0,
  x.start = NULL,
  n.sample = 1e+05,
  n.burnin = 10000,
  chain.id = 1,
  thinning = 1
)
}
\arguments{
\item{A}{Configuration Matrix}

\item{y}{(Corrupted/aggregated) Observation Vector}

\item{B}{Set of moves used (p-Markov, CPLB or Full Markov)}

\item{Model}{Type of model ("uniform" or "poisson")}

\item{lambda}{Rate parameters for "poisson" Model}

\item{p}{Relaxation limit}

\item{ldelta}{Combined normalizing constant and delta}

\item{w}{Weight put on "outsideness"}

\item{x.start}{Starting points for each chain. Needs to be fed in as a matrix with number of rows = number of chains}

\item{n.sample}{Number of samples}

\item{n.burnin}{Number of burnin steps}

\item{chain.id}{Which number chain?}

\item{thinning}{Thinning parameter}
}
\description{
p-samplers
}
