% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbmCov.R
\name{rbmCov}
\alias{rbmCov}
\title{Compute the multidimensional ESS for the sample}
\usage{
rbmCov(chains, batchSize = NULL)
}
\arguments{
\item{chains}{Sampled chains in list format (each item is a chain in matrix format without indices)}

\item{batchSize}{Variable used to evaluate the covarince matrix}
}
\description{
Compute the multidimensional ESS for the sample
}
