% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_draws_df.R
\name{to_draws_df}
\alias{to_draws_df}
\title{Converting from matrix to draws_df}
\usage{
to_draws_df(x, varNames = NULL, matrix = TRUE)
}
\arguments{
\item{x}{Matrix containing the sample including chain, iteration and draw numbers.}

\item{varNames}{Fill names for the variables if NULL names will be $xi$ for the ith variable.}

\item{matrix}{Is the sample a matrix?}
}
\value{
Returns the sample object of any of the samplers as a draws_df file
}
\description{
Converting from matrix to draws_df
}
