% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceplot.R
\name{traceplot}
\alias{traceplot}
\title{Plot traceplots for the variables}
\usage{
traceplot(x, variables)
}
\arguments{
\item{x}{The output from the MCMC should be put in.}

\item{variables}{Select a subset of variables whose traceplot should be plotted}
}
\value{
Plots a traceplot for each variable in the subset defined in variables
}
\description{
Plot traceplots for the variables
}
