% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_weights.R
\name{uniform_weights}
\alias{uniform_weights}
\title{Computes geometric weights for every move in B}
\usage{
uniform_weights(A, y, B)
}
\arguments{
\item{A}{Configuration matrix}

\item{y}{Observation vector}

\item{B}{Matrix containing basis used in the Hit-and-Run sampler. Every column is a move from the basis}
}
\value{
Returns a vector of weights used for the weighted_move_sampler
}
\description{
Computes geometric weights for every move in B
}
