% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_sampler.R
\name{x_sampler}
\alias{x_sampler}
\title{x-sampler}
\usage{
x_sampler(
  A,
  y,
  B,
  dist = "Unif",
  rate = NULL,
  proposal = "Random",
  n.sample = 1e+05,
  n.burnin = 10000,
  n.chains = 4,
  thinning = 1
)
}
\arguments{
\item{A}{Configuration matrix}

\item{y}{Vector of observations}

\item{B}{Full Markov bases to be used}

\item{dist}{Unconditional distribution on x}

\item{rate}{Poisson rates for each x_i}

\item{proposal}{Which proposal for moves? "Random" or "NonRandom", first will use the uniform distribution the latter a systematic approach.}

\item{n.sample}{Number of samples that are produced}

\item{n.burnin}{Number of burnin steps}

\item{n.chains}{Number of chains the sample is produced from}

\item{thinning}{Thinning parameter (store only every ith state)}
}
\value{
Returns a matrix of samples
}
\description{
x-sampler
}
