% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xSampler.R
\name{xSampler}
\alias{xSampler}
\title{x-sampler}
\usage{
xSampler(
  A,
  y,
  B,
  Model = "uniform",
  lambda = NULL,
  proposal = "Random",
  xStart = NULL,
  nSample = 1e+05,
  nBurnin = 10000,
  chainID = 1,
  thinning = 1,
  includeBurnin = FALSE
)
}
\arguments{
\item{A}{Configuration matrix}

\item{y}{Vector of observations}

\item{B}{Full Markov bases to be used}

\item{Model}{Unconditional distribution on x}

\item{lambda}{Poisson rates for each x_i}

\item{proposal}{Which proposal for moves? "Random" or "NonRandom", first will use the uniform distribution the latter a systematic approach.}

\item{xStart}{Starting points for each chain. Needs to be fed in as a matrix with number of rows = number of chains}

\item{nSample}{Number of samples that are produced}

\item{nBurnin}{Number of burnin steps}

\item{chainID}{Chain ID passed down from sampler.R}

\item{thinning}{Thinning parameter (store only every ith state)}

\item{includeBurnin}{Should the output include the burn-in samples?}
}
\value{
Returns a matrix of samples
}
\description{
x-sampler
}
