% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceptanceExt.R
\name{acceptanceExt}
\alias{acceptanceExt}
\title{Calculate acceptance probability}
\usage{
acceptanceExt(xCurrent, xProposal, loggamma, w, Model, lambda = NULL)
}
\arguments{
\item{xCurrent}{Current state of MCMC}

\item{xProposal}{Proposed new state of the MCMC}

\item{loggamma}{Combined normalizing constant and delta}

\item{w}{Weight put on "outsideness"}

\item{Model}{Type of model ("uniform" or "poisson")}

\item{lambda}{Rate parameters for "poisson" Model}
}
\description{
Calculate acceptance probability
}
