% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateAlpha.R
\name{calculateAlpha}
\alias{calculateAlpha}
\title{Calculate Acceptance ratio}
\usage{
calculateAlpha(x, x.star, ldelta, w = 0, dist = "Unif")
}
\arguments{
\item{x}{Current point}

\item{x.star}{Proposed point}

\item{ldelta}{log(delta) used to define epsilon}

\item{w}{Weight of negativity}

\item{dist}{Model distribution}
}
\value{
Returns the Acceptance ratio
}
\description{
Calculate Acceptance ratio
}
