% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperrectangleSampler.R
\name{hyperrectangleSampler}
\alias{hyperrectangleSampler}
\title{Extension Samplers}
\usage{
hyperrectangleSampler(
  A,
  y,
  B = NULL,
  Model = "uniform",
  lambda = NULL,
  A2Idx = NULL,
  a = NULL,
  loggamma = 0,
  w = 0,
  xStart = NULL,
  nSample = 1e+05,
  nBurnin = 10000,
  chainID = 4,
  thinning = 1
)
}
\arguments{
\item{A}{Configuration Matrix}

\item{y}{(Corrupted/aggregated) Observation Vector}

\item{B}{Set of moves used (p-Markov, CPLB or Full Markov)}

\item{Model}{Type of model ("uniform" or "poisson")}

\item{lambda}{Rate parameters for "poisson" Model}

\item{A2Idx}{Free variables correponding to matrix A_2}

\item{a}{"hyperrectangle" extension: bounds on hyperrectangle, "knapsack": objective function}

\item{loggamma}{Combined normalizing constant and delta}

\item{w}{Weight put on "outsideness"}

\item{xStart}{Starting points for each chain. Needs to be fed in as a matrix with number of rows = number of chains}

\item{nSample}{Number of samples}

\item{nBurnin}{Number of burnin steps}

\item{chainID}{Which number chain?}

\item{thinning}{Thinning parameter}
}
\description{
Extension Samplers
}
