% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposePoint.R
\name{proposePoint}
\alias{proposePoint}
\title{Proposing algorithm}
\usage{
proposePoint(
  x.current,
  moveIdx,
  B,
  extension = "p",
  p = 0,
  CPLBIdx = NULL,
  a = NULL,
  b = NULL
)
}
\arguments{
\item{x.current}{Current state of MCMC}

\item{moveIdx}{Index of the move in the collection of moves B}

\item{B}{Collection of moves (p-MB, CPLB or FMB)}

\item{extension}{Type of extension}

\item{p}{Relaxation limit}

\item{CPLBIdx}{Free variables correponding to matrix A_2}

\item{a}{"hyperrectangle" extension: bounds on hyperrectangle, "knapsack": objective function}

\item{b}{Upper bound of objective function for knapsack extension}
}
\description{
Proposing algorithm
}
