% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pSampler.R
\name{pSampler}
\alias{pSampler}
\title{p-samplers}
\usage{
pSampler(
  A,
  y,
  B,
  Model = "uniform",
  lambda = NULL,
  p = 0,
  loggamma = 0,
  w = 0,
  xStart = NULL,
  nSample = 1e+05,
  nBurnin = 10000,
  chainID = 1,
  thinning = 1
)
}
\arguments{
\item{A}{Configuration Matrix}

\item{y}{(Corrupted/aggregated) Observation Vector}

\item{B}{Set of moves used (p-Markov, CPLB or Full Markov)}

\item{Model}{Type of model ("uniform" or "poisson")}

\item{lambda}{Rate parameters for "poisson" Model}

\item{p}{Relaxation limit}

\item{loggamma}{Combined normalizing constant and delta}

\item{w}{Weight put on "outsideness"}

\item{xStart}{Starting points for each chain. Needs to be fed in as a matrix with number of rows = number of chains}

\item{nSample}{Number of samples}

\item{nBurnin}{Number of burnin steps}

\item{chainID}{Which number chain?}

\item{thinning}{Thinning parameter}
}
\description{
p-samplers
}
