% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_move_sample.R
\name{weighted_move_sampler}
\alias{weighted_move_sampler}
\title{Weighted moves sampler}
\usage{
weighted_move_sampler(
  A,
  y,
  B,
  dist = "Unif",
  rate = NULL,
  n.sample = 1e+05,
  n.burnin = 10000,
  n.chains = 4,
  thinning = 1,
  a = 1,
  weights = NULL
)
}
\arguments{
\item{A}{Configuration matrix}

\item{y}{Vector of observations}

\item{B}{Full Markov bases to be used}

\item{dist}{Unconditional distribution on x}

\item{rate}{Poisson rates for each x_i}

\item{n.sample}{Number of samples that are produced}

\item{n.burnin}{Number of burnin steps}

\item{n.chains}{Number of chains the sample is produced from}

\item{thinning}{Thinning parameter (store only every ith state)}

\item{a}{Tuning parameter determining how much heavier weights count towards sampling of moves}

\item{weights}{Weights for the individual moves. If NULL uses uniform weights. (Default - NULL)}
}
\value{
Returns a matrix of samples
}
\description{
Weighted moves sampler
}
