% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xSampler.R
\name{xSampler}
\alias{xSampler}
\title{x-sampler}
\usage{
xSampler(
  A,
  y,
  B,
  Model = "Unif",
  lambda = NULL,
  proposal = "Random",
  nSample = 1e+05,
  nBurnin = 10000,
  nChains = 4,
  thinning = 1
)
}
\arguments{
\item{A}{Configuration matrix}

\item{y}{Vector of observations}

\item{B}{Full Markov bases to be used}

\item{Model}{Unconditional distribution on x}

\item{lambda}{Poisson rates for each x_i}

\item{proposal}{Which proposal for moves? "Random" or "NonRandom", first will use the uniform distribution the latter a systematic approach.}

\item{nSample}{Number of samples that are produced}

\item{nBurnin}{Number of burnin steps}

\item{nChains}{Number of chains the sample is produced from}

\item{thinning}{Thinning parameter (store only every ith state)}
}
\value{
Returns a matrix of samples
}
\description{
x-sampler
}
